requirements.txt
# Core dependencies
pandas>=1.3.0
requests>=2.26.0
python-dotenv>=0.19.0
matplotlib>=3.4.0
seaborn>=0.11.0
numpy>=1.21.0
openpyxl>=3.0.7  # For Excel support

# Data processing
scipy>=1.7.0
statsmodels>=0.13.0  # For time series analysis

# Interactive exploration
inquirer>=2.7.0  # For interactive CLI

# Documentation
jupyter>=1.0.0  # For notebooks

# Development
twine>=3.4.2
wheel>=0.37.0
pytest>=6.2.5
